added option to automatically keep your phoebe installation up-to-date. Suppose you have a computer without the Phoebe svn downloaded or installed, then the only thing you need to do is download the install.py script, and type in a terminal: "python install.py --auto-update my_installation_directory", and everything will be downloaded and installed automatically. That is, all the dependencies (numpy, matplotlib, pyfits, emcee...), the SVN and the atmosphere files. All the Fortran and C libraries will be compiled and installed. If my_installation_directory already exists, the current installation will be updated (not completely reinstalled). Adding the --auto-update executes the script each day at midnight, keeping the installation up-to-date. Uninstalling is simply done removing the directory. Activating the installation (overriding your default python installation) is done via "source my_installation_directory/bin/activate", deactivating your installation is then simply done by typing "deactivate". You can happily install different versions simultaneously.

+++ devel/install.py
@@ -1,9 +1,10 @@
 """
 Install Phoebe and all its dependencies in a virtual environment.
 
-To install, simply do::
+To install, simply do one of::
 
     $:> python install.py MYDIR
+    $:> python install.py --auto-update MYDIR
     
 With ``MYDIR`` the directory where Phoebe needs to be installed to. ``MYDIR``
 defaults to the directory ``phoebe`` in the current working directory. Make sure
@@ -12,6 +13,15 @@ the directory does not exist.
 To update, simply do the same as for the installation::
 
     $:> python install.py MYDIR
+    
+If you add the option ``--auto-update`` during install or update, the
+installation will be automatically updated at midnight every day. In that case,
+make sure the ``install.py`` file remains at the same location. Removing the
+cronjob can be done via::
+
+    $:> crontab -e
+    
+and following the instructions.
     
 To unistall, simply remove the directory::
 
@@ -37,9 +47,16 @@ def line_at_a_time(fileobj):
         if not line:
             return
         yield line
+
+if '--auto-update' in sys.argv[1:]:
+    add_cronjob = tuple(list(sys.argv).pop(sys.argv.index('--auto-update')))
+else:
+    add_cronjob = False
         
 if len(sys.argv)==1:
     sys.argv.append('phoebe')
+    
+
 
 base_dir = sys.argv[-1]
 
@@ -58,9 +75,11 @@ else:
     virtualenv.main()
 
 #-- download the requirement files
-urllib.urlretrieve('http://www.phoebe-project.org/2.0/docs/_downloads/numpy-basic.txt', "numpy-basic.txt")
-urllib.urlretrieve('http://www.phoebe-project.org/2.0/docs/_downloads/phoebe-basic.txt', "phoebe-basic.txt")
-urllib.urlretrieve('http://www.phoebe-project.org/2.0/docs/_downloads/phoebe-full.txt', "phoebe-full.txt")
+req_files = ['numpy-basic.txt','phoebe-basic.txt','phoebe-full.txt']
+for req_file in req_files:
+    if not os.path.isfile(req_file):
+        print("Downloaded requirement file {}".format(req_file))
+        urllib.urlretrieve('http://www.phoebe-project.org/2.0/docs/_downloads/{}'.format(req_file), req_file)
 
 #-- install the basic requirements
 activate = '. '+os.path.join(base_dir,'bin','activate')
@@ -71,6 +90,7 @@ fail = False
 
 new_package = False
 nr = 0
+last_char_was_dot = False
 with open('install.log','w') as log:
     for thing_to_do in things_to_do:
         cmd = ' '.join([activate]+['&&']+thing_to_do)
@@ -79,17 +99,22 @@ with open('install.log','w') as log:
         for line in line_at_a_time(p1.stdout):
             log.write(line)
             if 'Requirement already satisfied' in line:
-                print('')
+                if last_char_was_dot:
+                    print('')
+                    last_char_was_dot = False
                 print(line.strip())
             elif 'Successfully installed' in line:
                 new_package = False
                 sys.stdout.write('done.\n')
                 sys.stdout.flush()
                 print(line.strip())
+                last_char_was_dot = False
             elif "Downloading/unpacking" in line or 'Obtaining' in line:
                 new_package = True
                 line = line.split()
-                print('')
+                if last_char_was_dot:
+                    print('')
+                    last_char_was_dot = False
                 sys.stdout.write(' '.join(line[:1]+['/installing']+line[1:]))
                 sys.stdout.flush()
             elif new_package:
@@ -97,7 +122,11 @@ with open('install.log','w') as log:
                 if nr%100==0:
                     sys.stdout.write('.')
                     sys.stdout.flush()
-        print('')
+                    last_char_was_dot = True
+        if last_char_was_dot:
+            print('')
+            last_char_was_dot = False
+                
         log.flush()
 
 #-- finally download the atmosphere files:
@@ -106,5 +135,23 @@ files = ['kurucz_p00_claret_equidist_r_leastsq_teff_logg.fits',
 
 for ff in files:
     source = 'http://www.phoebe-project.org/2.0/docs/_downloads/'+ff
-    destin = os.path.join(basedir,'src/phoebe/atmospheres/tables/ld_coeffs/'+ff)
-    urllib.urlretrieve(source,destin)
+    destin = os.path.join(base_dir,'src/phoebe/phoebe/atmospheres/tables/ld_coeffs/'+ff)
+    if not os.path.isfile(destin):
+        urllib.urlretrieve(source,destin)
+        print("Added table {}".format(destin))
+    else:
+        print("Table {} already exists".format(destin))
+
+#-- if you want a Cron job; do that here:
+if add_cronjob:
+    cronjob = "0 * * * * python {} {}".format(os.path.abspath(sys.argv[0]),base_dir)
+    flag = subprocess.call('(crontab -l ; echo "{}") |uniq - | crontab -'.format(cronjob),shell=True)
+    if not flag:
+        print("Succesfully added cronjob '{}'".format(cronjob))
+    else:
+        print("Could not add cronjob")
+else:
+    print("Did not add cronjob for regular updating")
+
+print("Don't forget to source the virtual environment before using Phoebe:\nsource {}".format(os.path.join(base_dir,'bin/activate')))
+
